PhotoImage (test1)
i trial this many times on another replit with new photos but i can never seem to get it to work like my first version (ping pong the game) back 3 years ago. so i've decided to try out an image from my first game.
although the images is not quite to size with the frame